+++ modules/elementary_functions/src/c/dllMain.c
@@ -24,8 +24,12 @@
 // LICENCE_BLOCK_END
 //=============================================================================
 #include <Windows.h>
+#include "nlsConfig.h"
 //=============================================================================
 #pragma comment(lib, "libnlsblaslapack.lib")
+#if defined(_NLS_WITH_VML)
+#pragma comment(lib, "libnlsvml_mkl.lib")
+#endif
 //=============================================================================
 int WINAPI
 DllMain(HINSTANCE hInstance, DWORD reason, PVOID pvReserved)