+++ modules/elementary_functions/src/cpp/TruncateFunctions.cpp
@@ -23,15 +23,18 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include <cmath>
-#include <functional>
-#include <Eigen/Dense>
 #include "nlsConfig.h"
+#undef _NLS_WITH_VML
+#include <cmath>
+#if defined(_NLS_WITH_VML)
+#include <mkl_vml.h>
+#endif
 #include "TruncateFunctions.hpp"
 #include "ClassName.hpp"
 #include "characters_encoding.hpp"
 //=============================================================================
 namespace Nelson {
+//=============================================================================
 typedef enum
 {
     ROUND,
@@ -40,181 +43,355 @@ typedef enum
     FLOOR
 } TRUNCATE_LEVEL;
 //=============================================================================
-static double
-fixDouble(double a)
+template <class T>
+static void
+oTrunc(indexType len, const T* pIn, T* pOut)
 {
-    return std::trunc(a);
-}
-//=============================================================================
-static single
-fixSingle(single a)
-{
-    return std::trunc(a);
-}
-//=============================================================================
-static double
-ceilDouble(double a)
-{
-    return ceil(a);
-}
-//=============================================================================
-static single
-ceilSingle(single a)
-{
-    return ceil(a);
-}
-//=============================================================================
-static double
-roundDouble(double a)
-{
-    return round(a);
-}
-//=============================================================================
-static single
-roundSingle(single a)
-{
-    return round(a);
-}
-//=============================================================================
-static double
-floorDouble(double a)
-{
-    return floor(a);
-}
-//=============================================================================
-static single
-floorSingle(single a)
-{
-    return floor(a);
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+    for (ompIndexType i = 0; i < (ompIndexType)(len); i++) {
+        if (std::isfinite((T)pIn[i])) {
+            pOut[i] = std::trunc((T)pIn[i]);
+        } else {
+            pOut[i] = (T)pIn[i];
+        }
+    }
 }
 //=============================================================================
 template <class T>
-ArrayOf
-truncateArray(ArrayOf arrayIn, T (*ptrFunc)(T))
+static void
+oRound(indexType len, const T* pIn, T* pOut)
 {
-    size_t len = arrayIn.getElementCount();
-    void* ptr = ArrayOf::allocateArrayOf(arrayIn.getDataClass(), len, stringVector(), true);
-    T* rp = (T*)ptr;
-    T* dp = (T*)arrayIn.getDataPointer();
-
 #if defined(_NLS_WITH_OPENMP)
 #pragma omp parallel for
 #endif
-    for (ompIndexType i = 0; i < (ompIndexType)(len * 2); i++) {
-        if (std::isfinite(dp[i])) {
-            rp[i] = (ptrFunc)((T)dp[i]);
+    for (ompIndexType i = 0; i < (ompIndexType)(len); i++) {
+        if (std::isfinite((T)pIn[i])) {
+            pOut[i] = std::round((T)pIn[i]);
         } else {
-            rp[i] = dp[i];
+            pOut[i] = (T)pIn[i];
         }
     }
-    return ArrayOf(arrayIn.getDataClass(), arrayIn.getDimensions(), rp);
+}
+//=============================================================================
+template <class T>
+static void
+oFloor(indexType len, const T* pIn, T* pOut)
+{
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+    for (ompIndexType i = 0; i < (ompIndexType)(len); i++) {
+        if (std::isfinite((T)pIn[i])) {
+            pOut[i] = std::floor((T)pIn[i]);
+        } else {
+            pOut[i] = (T)pIn[i];
+        }
+    }
+}
+//=============================================================================
+template <class T>
+static void
+oCeil(indexType len, const T* pIn, T* pOut)
+{
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+    for (ompIndexType i = 0; i < (ompIndexType)(len); i++) {
+        if (std::isfinite((T)pIn[i])) {
+            pOut[i] = std::ceil((T)pIn[i]);
+        } else {
+            pOut[i] = (T)pIn[i];
+        }
+    }
 }
 //=============================================================================
 static ArrayOf
-Truncate(ArrayOf arrayIn, TRUNCATE_LEVEL level)
+FixSingleComplex(const ArrayOf& arrayIn)
 {
-    ArrayOf res;
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsTrunc((MKL_INT)(len * 2), dp, ptr);
+#else
+    oTrunc<single>(len * 2, dp, ptr);
+#endif
+    return ArrayOf(NLS_SCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FixDoubleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdTrunc((MKL_INT)(len * 2), dp, ptr);
+#else
+    oTrunc<double>(len * 2, dp, ptr);
+#endif
+    return ArrayOf(NLS_DCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FixSingle(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsTrunc((MKL_INT)len, dp, ptr);
+#else
+    oTrunc<single>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_SINGLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FixDouble(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdTrunc((MKL_INT)len, dp, ptr);
+#else
+    oTrunc<double>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+RoundDoubleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdRound((MKL_INT)(len * 2), dp, ptr);
+#else
+    oRound<double>((len * 2), dp, ptr);
+#endif
+    return ArrayOf(NLS_DCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+RoundSingleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsRound((MKL_INT)(len * 2), dp, ptr);
+#else
+    oRound<single>((len * 2), dp, ptr);
+#endif
+    return ArrayOf(NLS_SCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+RoundDouble(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdRound((MKL_INT)len, dp, ptr);
+#else
+    oRound<double>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+RoundSingle(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsRound((MKL_INT)len, dp, ptr);
+#else
+    oRound<single>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_SINGLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FloorSingleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsFloor((MKL_INT)(len * 2), dp, ptr);
+#else
+    oFloor<single>(len * 2, dp, ptr);
+#endif
+    return ArrayOf(NLS_SCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FloorDoubleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdFloor((MKL_INT)(len * 2), dp, ptr);
+#else
+    oFloor<double>(len * 2, dp, ptr);
+#endif
+    return ArrayOf(NLS_DCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FloorSingle(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsFloor((MKL_INT)len, dp, ptr);
+#else
+    oFloor<single>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_SINGLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+FloorDouble(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdFloor((MKL_INT)len, dp, ptr);
+#else
+    oFloor<double>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+CeilSingleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsCeil((MKL_INT)(len * 2), dp, ptr);
+#else
+    oCeil<single>((len * 2), dp, ptr);
+#endif
+    return ArrayOf(NLS_SCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+CeilDoubleComplex(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdCeil((MKL_INT)(len * 2), dp, ptr);
+#else
+    oCeil<double>((len * 2), dp, ptr);
+#endif
+    return ArrayOf(NLS_DCOMPLEX, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+CeilSingle(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, len, stringVector(), false);
+    const single* dp = (const single*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vsCeil((MKL_INT)len, dp, ptr);
+#else
+    oCeil<single>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_SINGLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static ArrayOf
+CeilDouble(const ArrayOf& arrayIn)
+{
+    indexType len = arrayIn.getElementCount();
+    double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
+    const double* dp = (const double*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_VML)
+    vdCeil((MKL_INT)len, dp, ptr);
+#else
+    oCeil<double>(len, dp, ptr);
+#endif
+    return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), ptr);
+}
+//=============================================================================
+static std::wstring
+getNotSupportedTypeMessage(ArrayOf arrayIn, TRUNCATE_LEVEL level)
+{
+    std::wstring name;
+    ClassName(arrayIn, name);
+    std::wstring functionName;
+    switch (level) {
+    case TRUNCATE_LEVEL::CEIL:
+        functionName = L"ceil";
+        break;
+    case TRUNCATE_LEVEL::ROUND:
+        functionName = L"round";
+        break;
+    case TRUNCATE_LEVEL::FIX:
+        functionName = L"fix";
+        break;
+    case TRUNCATE_LEVEL::FLOOR:
+        functionName = L"floor";
+        break;
+    }
+    return _W("Undefined function '") + name + L"_" + functionName + L"'";
+}
+//=============================================================================
+static ArrayOf
+Truncate(const ArrayOf &arrayIn, TRUNCATE_LEVEL level)
+{
     if (arrayIn.isSparse()) {
-        switch (level) {
-        case TRUNCATE_LEVEL::CEIL:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"ceil"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::ROUND:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"round"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::FIX:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"fix"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::FLOOR:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"floor"
-                + L"'");
-            break;
-        }
+        Error(getNotSupportedTypeMessage(arrayIn, level));
     }
     switch (arrayIn.getDataClass()) {
     case NLS_SCOMPLEX: {
-        single (*ptr)(single);
         switch (level) {
         case TRUNCATE_LEVEL::CEIL: {
-            ptr = &(ceilSingle);
-            return truncateArray<single>(arrayIn, ptr);
+            return CeilSingleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::ROUND: {
-            ptr = &(roundSingle);
-            return truncateArray<single>(arrayIn, ptr);
+            return RoundSingleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::FIX: {
-            ptr = &(fixSingle);
-            return truncateArray<single>(arrayIn, ptr);
+            return FixSingleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::FLOOR: {
-            ptr = &(floorSingle);
-            return truncateArray<single>(arrayIn, ptr);
+            return FloorSingleComplex(arrayIn);
         } break;
         }
-        return res;
     } break;
     case NLS_DCOMPLEX: {
-        double (*ptr)(double);
         switch (level) {
         case TRUNCATE_LEVEL::CEIL: {
-            ptr = &(ceilDouble);
-            return truncateArray<double>(arrayIn, ptr);
+            return CeilDoubleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::ROUND: {
-            ptr = &(roundDouble);
-            return truncateArray<double>(arrayIn, ptr);
+            return RoundDoubleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::FIX: {
-            ptr = &(fixDouble);
-            return truncateArray<double>(arrayIn, ptr);
+            return FixDoubleComplex(arrayIn);
         } break;
         case TRUNCATE_LEVEL::FLOOR: {
-            ptr = &(floorDouble);
-            return truncateArray<double>(arrayIn, ptr);
+            return FloorDoubleComplex(arrayIn);
         } break;
         }
-    } break;
-    case NLS_HANDLE:
-    case NLS_GO_HANDLE:
-    case NLS_CELL_ARRAY:
-    case NLS_STRING_ARRAY:
-    case NLS_STRUCT_ARRAY:
-    default: {
-        switch (level) {
-        case TRUNCATE_LEVEL::CEIL:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"ceil"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::ROUND:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"round"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::FIX:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"fix"
-                + L"'");
-            break;
-        case TRUNCATE_LEVEL::FLOOR:
-            Error(_W("Undefined function '") + utf8_to_wstring(ClassName(arrayIn)) + L"_" + L"floor"
-                + L"'");
-            break;
-        }
-    } break;
-    case NLS_CHAR: {
-        size_t len = arrayIn.getElementCount();
-        void* ptr = ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
-        auto* rp = static_cast<double*>(ptr);
-        auto* dp = (charType*)arrayIn.getDataPointer();
-#if defined(_NLS_WITH_OPENMP)
-#pragma omp parallel for
-#endif
-        for (ompIndexType i = 0; i < (ompIndexType)len; i++) {
-            rp[i] = static_cast<double>(dp[i]);
-        }
-        res = ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), rp);
     } break;
     case NLS_LOGICAL: {
         size_t len = arrayIn.getElementCount();
@@ -227,9 +404,40 @@ Truncate(ArrayOf arrayIn, TRUNCATE_LEVEL level)
         for (ompIndexType i = 0; i < (ompIndexType)len; i++) {
             rp[i] = (dp[i] == 0 ? 0 : 1);
         }
-        res = ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), rp);
+        return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), rp);
     } break;
-    case NLS_DOUBLE:
+    case NLS_SINGLE: {
+        switch (level) {
+        case TRUNCATE_LEVEL::CEIL: {
+            return CeilSingle(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::ROUND: {
+            return RoundSingle(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::FIX: {
+            return FixSingle(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::FLOOR: {
+            return FloorSingle(arrayIn);
+        } break;
+        }
+    } break;
+    case NLS_DOUBLE: {
+        switch (level) {
+        case TRUNCATE_LEVEL::CEIL: {
+            return CeilDouble(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::ROUND: {
+            return RoundDouble(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::FIX: {
+            return FixDouble(arrayIn);
+        } break;
+        case TRUNCATE_LEVEL::FLOOR: {
+            return FloorDouble(arrayIn);
+        } break;
+        }
+    }
     case NLS_UINT8:
     case NLS_INT8:
     case NLS_UINT16:
@@ -237,76 +445,54 @@ Truncate(ArrayOf arrayIn, TRUNCATE_LEVEL level)
     case NLS_UINT32:
     case NLS_INT32:
     case NLS_UINT64:
-    case NLS_INT64:
-    case NLS_SINGLE: {
-        ArrayOf InputAsDouble(arrayIn);
-        InputAsDouble.ensureSingleOwner();
-        InputAsDouble.promoteType(NLS_DOUBLE);
-        ArrayOf OutputAsDouble(arrayIn);
-        OutputAsDouble.ensureSingleOwner();
-        OutputAsDouble.promoteType(NLS_DOUBLE);
-        switch (level) {
-        case TRUNCATE_LEVEL::CEIL: {
-            // to speed up computations, we use a vector with eigen library and MKL
-            Eigen::Map<Eigen::MatrixXd> matA(
-                (double*)InputAsDouble.getDataPointer(), 1, InputAsDouble.getElementCount());
-            Eigen::Map<Eigen::MatrixXd> matR(
-                (double*)OutputAsDouble.getDataPointer(), 1, OutputAsDouble.getElementCount());
-            matR = matA.array().ceil();
-        } break;
-        case TRUNCATE_LEVEL::ROUND: {
-            // to speed up computations, we use a vector with eigen library and MKL
-            Eigen::Map<Eigen::MatrixXd> matA(
-                (double*)InputAsDouble.getDataPointer(), 1, InputAsDouble.getElementCount());
-            Eigen::Map<Eigen::MatrixXd> matR(
-                (double*)OutputAsDouble.getDataPointer(), 1, OutputAsDouble.getElementCount());
-            matR = matA.array().round();
-
-        } break;
-        case TRUNCATE_LEVEL::FIX: {
-            Eigen::Map<Eigen::MatrixXd> matA(
-                (double*)InputAsDouble.getDataPointer(), 1, InputAsDouble.getElementCount());
-            Eigen::Map<Eigen::MatrixXd> matR(
-                (double*)OutputAsDouble.getDataPointer(), 1, OutputAsDouble.getElementCount());
-            matR = matA.unaryExpr(std::ref(fixDouble));
-
-        } break;
-        case TRUNCATE_LEVEL::FLOOR: {
-            // to speed up computations, we use a vector with eigen library and MKL
-            Eigen::Map<Eigen::MatrixXd> matA(
-                (double*)InputAsDouble.getDataPointer(), 1, InputAsDouble.getElementCount());
-            Eigen::Map<Eigen::MatrixXd> matR(
-                (double*)OutputAsDouble.getDataPointer(), 1, OutputAsDouble.getElementCount());
-            matR = matA.array().floor();
-        } break;
+    case NLS_INT64: {
+        return arrayIn;
+    } break;
+    case NLS_CHAR: {
+        size_t len = arrayIn.getElementCount();
+        void* ptr = ArrayOf::allocateArrayOf(NLS_DOUBLE, len, stringVector(), false);
+        auto* rp = static_cast<double*>(ptr);
+        auto* dp = (charType*)arrayIn.getDataPointer();
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+        for (ompIndexType i = 0; i < (ompIndexType)len; i++) {
+            rp[i] = static_cast<double>(dp[i]);
         }
-        OutputAsDouble.promoteType(arrayIn.getDataClass());
-        return OutputAsDouble;
+        return ArrayOf(NLS_DOUBLE, arrayIn.getDimensions(), rp);
+    } break;
+    case NLS_HANDLE:
+    case NLS_GO_HANDLE:
+    case NLS_CELL_ARRAY:
+    case NLS_STRING_ARRAY:
+    case NLS_STRUCT_ARRAY:
+    default: {
+        Error(getNotSupportedTypeMessage(arrayIn, level));
     } break;
     }
-    return res;
+    return ArrayOf();
 }
 //=============================================================================
 ArrayOf
-Round(ArrayOf arrayIn)
+Round(const ArrayOf &arrayIn)
 {
     return Truncate(arrayIn, TRUNCATE_LEVEL::ROUND);
 }
 //=============================================================================
 ArrayOf
-Ceil(ArrayOf arrayIn)
+Ceil(const ArrayOf& arrayIn)
 {
     return Truncate(arrayIn, TRUNCATE_LEVEL::CEIL);
 }
 //=============================================================================
 ArrayOf
-Floor(ArrayOf arrayIn)
+Floor(const ArrayOf& arrayIn)
 {
     return Truncate(arrayIn, TRUNCATE_LEVEL::FLOOR);
 }
 //=============================================================================
 ArrayOf
-Fix(ArrayOf arrayIn)
+Fix(const ArrayOf& arrayIn)
 {
     return Truncate(arrayIn, TRUNCATE_LEVEL::FIX);
 }